Move vote retrieval logic to the parent component
